refactor: ModalContainer 내부 useEffect 훅으로 분리
