+++ components/src/modal/hooks/useBodyScrollLock.ts
@@ -1,0 +1,15 @@
+import { useEffect } from 'react';
+
+export function useBodyScrollLock(open: boolean) {
+  useEffect(() => {
+    if (open) {
+      document.body.style.overflow = 'hidden';
+    } else {
+      document.body.style.overflow = '';
+    }
+
+    return () => {
+      document.body.style.overflow = '';
+    };
+  }, [open]);
+}
